Added option to prevent cache busting
Added bustCache option (and defaulted it to false) for controlling
whether or not images are requested with a unique query parameter
(effectively preventing client side caching). The default behavior prior
to this change was to bust the cache; this update changes the default
behavior to no longer bust the cache.